requirements.txt
appdirs==1.4.3
Django==1.10.6
django-filter==1.0.2
djangorestframework==3.6.2
packaging==16.8
pyparsing==2.2.0
six==1.10.0
twilio==6.1.0
jsonfield==2.0.1
django-dirtyfields==1.2.1
